Refactor sep24 to use shared util classes
