+++ cases-SEP24/fee.optional.test.js
@@ -1,10 +1,10 @@
 import StellarSDK from "stellar-sdk";
 
-import { fetch } from "./util/fetchShim";
+import { fetch } from "../util/fetchShim";
 import getTomlFile from "./util/getTomlFile";
 import getSep10Token from "./util/sep10";
 import { errorSchema, feeSchema } from "./util/schema";
-import { ensureCORS } from "./util/ensureCORS";
+import { ensureCORS } from "../util/ensureCORS";
 
 const keyPair = StellarSDK.Keypair.random();
 const urlBuilder = new URL(process.env.DOMAIN);
