+++ cases-SEP24/interactive-flows.optional.test.js
@@ -1,7 +1,7 @@
 /**
  * @jest-environment ./cases-SEP24/environment.js
  */
-import { fetch } from "./util/fetchShim";
+import { fetch } from "../util/fetchShim";
 import getSep10Token from "./util/sep10";
 import StellarSDK from "stellar-sdk";
 import getTomlFile from "./util/getTomlFile";
