--- cases-SEP24/util/fetchShim.js
@@ -1,4 +1,0 @@
-import doFetch from "node-fetch";
-export const fetch = (...args) => {
-  return doFetch(...args);
-};
