+++ cases-SEP24/util/friendbot.js
@@ -1,4 +1,4 @@
-import { fetch } from "./fetchShim";
+import { fetch } from "../../util/fetchShim";
 import StellarSdk from "stellar-sdk";
 
 const friends = [];
