+++ cases-SEP24/util/interactive.js
@@ -1,4 +1,4 @@
-import { fetch } from "./fetchShim";
+import { fetch } from "../../util/fetchShim";
 import { waitForLoad, openObservableWindow } from "./browser-util";
 import { getTransactionSchema } from "./schema";
 import FormData from "form-data";
