+++ server/jest-controller/run-test.js
@@ -6,9 +6,7 @@ module.exports = async (domain, currency, test) => {
   return new Promise((resolve, reject) => {
     const env = { ...process.env };
     env.DOMAIN = domain;
-    console.log("DOMAIN ", domain);
     env.CURRENCY = currency;
-    console.log("RUN TEST ", test);
     const jest = spawn(
       "node_modules/.bin/jest",
       ["--json", "--noStackTrace", `cases-${PROJECT}/${test}.test.js`],
